requirements.txt
# Audio Visualizer Pro - Requirements
# Python 3.8+ required

# Audio-Analyse
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0

# Bildverarbeitung
Pillow>=9.0.0

# Datenvalidierung
pydantic>=1.10.0

# CLI
click>=8.0.0

# Fortschrittsbalken (optional, für bessere UX)
tqdm>=4.62.0

# Testing
pytest>=7.0.0

# Audio-I/O (für Test-Dateien)
soundfile>=0.11.0

# GUI
streamlit>=1.28.0

# FFmpeg wird system-seitig benötigt:
# Ubuntu/Debian: sudo apt-get install ffmpeg
# macOS: brew install ffmpeg
# Windows: https://ffmpeg.org/download.html
